requirements.txt
flask
pywebview
gunicorn
pyspeedtest
apscheduler